pyproject.toml
[project]
name = "sk-hynix-multi-agent-prototype"
version = "0.1.0"
description = "Multi-service FastAPI + LangGraph prototype for SK hynix future market direction analysis."
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi==0.115.12",
    "uvicorn==0.34.2",
    "langgraph==0.4.7",
    "httpx==0.28.1",
    "pydantic==2.11.3",
    "python-dotenv==1.1.0",
    "Pillow==11.2.1",
    "reportlab==4.4.1",
    "openai>=1.0.0,<2.0.0",
    "pypdf==5.4.0",
    "qdrant-client==1.14.2",
    "sentence-transformers==3.4.1",
]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
